+++ CHANGELOG.md
@@ -5,6 +5,7 @@ omnibus-gitlab repository.
 
 11.8.0
 
+- Remove Redis config from gitlab-shell !3000
 - Add AWS customer provided encryption key configuration option !2928
 - Fix mattermost data directory permissions with ```update-permissions```
 - Patch runit to not consider status of log service for `status` command exit